Fix #5565
After reading the kart version from kart.xml, we immediately abort all kart properties loading operations if the kart version is not supported by the binary. This prevents the OOB error from occurring.

Care is taken to show only a warning message about the kart not being loaded instead of error messages.

An additional check that the minimum kart version is inferior or equal to the maximum kart version has also been added.